reorganize repo to clean out top-level dir
