+++ agents/BnnRTabularQ.py
@@ -1,6 +1,6 @@
 import numpy as np
 from agents.TabularQ import TabularQ
-from BNNApproximation import BNNApproximation
+from utils.BNNApproximation import BNNApproximation
 
 class BnnRTabularQ(TabularQ):
     def __init__(self, state_shape, num_acts, params):
