+++ agents/TabularRTabularQ.py
@@ -1,6 +1,6 @@
 import numpy as np
 from agents.TabularQ import TabularQ
-from bayesianapproximator import TabularBayesianApproximation
+from utils.bayesianapproximator import TabularBayesianApproximation
 
 class TabularRTabularQ(TabularQ):
     def __init__(self, state_shape, num_acts, params):
