+++ build.sh
@@ -1,5 +1,5 @@
-cd sparse
-python3 setup.py install --home=../
-cd ..
+cd utils/sparse
+python3 setup.py install --home=../../
+cd ../..
 touch lib/__init__.py
 touch lib/python/__init__.py
