+++ scripts/plot.py
