+++ utils/sparse/sparse_module.c
